+++ src/Services/Http/RetrieverHttpClientFactory.php
@@ -2,62 +2,27 @@
 
 namespace App\Services\Http;
 
-use GuzzleHttp\Client;
 use GuzzleHttp\Cookie\CookieJarInterface;
 use GuzzleHttp\HandlerStack;
 
-class RetrieverHttpClientFactory
+class RetrieverHttpClientFactory extends HttpClientFactory
 {
-    /**
-     * @var array
-     */
-    private $curlOptions;
-
     /**
      * @var CookieJarInterface
      */
     private $cookieJar;
 
-    /**
-     * @var HandlerStack
-     */
-    private $handlerStack;
-
     public function __construct(array $curlOptions, HandlerStack $handlerStack, CookieJarInterface $cookieJar)
     {
-        $this->setCurlOptions($curlOptions);
+        parent::__construct($curlOptions, $handlerStack);
 
         $this->cookieJar = $cookieJar;
-        $this->handlerStack = $handlerStack;
     }
 
-    /**
-     * @return Client
-     */
-    public function create()
+    protected function createClientConfig(): array
     {
-        return new Client([
-            'curl' => $this->curlOptions,
-            'verify' => false,
-            'handler' => $this->handlerStack,
-            'max_retries' => HttpRetryMiddlewareFactory::MAX_RETRIES,
+        return array_merge(parent::createClientConfig(), [
             'cookies' => $this->cookieJar,
         ]);
-    }
-
-    /**
-     * @param array $curlOptions
-     */
-    private function setCurlOptions(array $curlOptions)
-    {
-        $definedCurlOptions = [];
-
-        foreach ($curlOptions as $name => $value) {
-            if (defined($name)) {
-                $definedCurlOptions[constant($name)] = $value;
-            }
-        }
-
-        $this->curlOptions = $definedCurlOptions;
     }
 }
